use say to send messages in another channel

+++ src/commands/say.ts
@@ -1,5 +1,7 @@
+import { NewsChannel, TextChannel } from "discord.js";
 import { Command } from "src/gm";
 import { permLevels } from "../permissions";
+import { stringToChannel } from "../utils/parsers";
 import xlg from "../xlogger";
 
 export const command: Command = {
@@ -9,15 +11,26 @@ export const command: Command = {
         long: "Make the bot say something. Must be moderator."
     },
     args: true,
-    usage: "<bot message>",
+    usage: "[channel] <bot message>",
     guildOnly: true,
     permLevel: permLevels.mod,
     category: 'utility',
     cooldown: 1,
     async execute(client, message, args) {
         try {
-            message.delete();
-            message.channel.send(args.join(" "));
+            if (!message.guild) return;
+            const channel = stringToChannel(message.guild, args[0], false, false);
+            if (channel && (channel instanceof TextChannel || channel instanceof NewsChannel)) {
+                args.shift();
+                if (!args.length) {
+                    client.specials?.sendError(message.channel, `Stuff to send was not specified.`);
+                    return;
+                }
+                await channel.send(args.join(" "));
+                return;
+            }
+            await message.delete();
+            await message.channel.send(args.join(" "));
         } catch (error) {
             xlg.error(error);
             await client.specials?.sendError(message.channel);
